Create README.md
Implementation of CRUD in the the project. It also uses the session too check whether the authenticate user is modifying the website or not. If, there is no authenticate user, it will kick you out to the login page. All the pages that the admin can use are secured using this feature.